refactor monitor
