+++ philo/srcs/actors/monitor/monitor_on_start.c
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   monitor_on_start.c                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/10 15:11:18 by ttsubo            #+#    #+#             */
+/*   Updated: 2025/06/21 18:45:10 by ttsubo           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "monitor_actor.h"
+
+void	monitor_on_start(t_ft_actor *self)
+{
+	self->parent->tell(self->parent, msg_new(ACTOR_START_DONE, self, NULL));
+}
